requirements.txt
google-adk==1.13
uvicorn
fastapi
google-genai